CF2-014: refactor: Move origin-related requests to the store and reorganize components for better structure
