+++ src/components/Form/Origin/hooks/useOriginForm.hook.ts
@@ -1,9 +1,9 @@
 import type { Origin } from "@/entities/origin.entity";
 import { useAppSearchParams } from "@/hooks/useAppSearchParams";
-import { useFindOriginById } from "@/requests/origin/useFindOriginById.request";
-import { useOriginCreate } from "@/requests/origin/useOriginCreate.request";
-import { useOriginUpdate } from "@/requests/origin/useOriginUpdate.request";
 import { useAppDispatch, useAppSelector } from "@/store";
+import { useFindOriginById } from "@/store/requests/origin/useFindOriginById.request";
+import { useOriginCreate } from "@/store/requests/origin/useOriginCreate.request";
+import { useOriginUpdate } from "@/store/requests/origin/useOriginUpdate.request";
 import { ShowAndHideActions } from "@/store/slices/showAndHide/showAndHide.slice";
 import { zodResolver } from "@hookform/resolvers/zod";
 import type { Dispatch, UnknownAction } from "@reduxjs/toolkit";
